+++ aim/sdk/__init__.py
@@ -1,16 +1,15 @@
 from aim.sdk.track import track, set_params
 from aim.sdk.init import init
 from aim.sdk.select import select
+from aim.sdk.session import Session
 
-
-# Serializable classes references
-metric = 'Metric'
-hyperparams = 'HyperParameters'
-dataset = 'Dataset'
-map = 'Map'
-nested_map = 'NestedMap'
-checkpoint = 'Checkpoint'
-
-# Serializable aliases
-params = hyperparams
-dictionary = nested_map
+from aim.artifacts import (
+    metric,
+    hyperparams,
+    params,
+    dataset,
+    dictionary,
+    map,
+    nested_map,
+    checkpoint,
+)